requirements.txt
pytorch-lightning==1.4.9
torch==1.9.0
torchvision==0.10.0
pandas==1.3.3
Pillow==8.2.0
scikit-learn==0.24.2
